requirements.txt
asgiref==3.5.2
Django==4.0.8
psycopg2-binary==2.9.3
sqlparse==0.4.3
tzdata==2022.6
django-allauth==0.50.0
django-environ==0.9.0
pillow==9.0.1
django-debug-toolbar==3.8.1

